Cleaned up how API requests are handled, added validity checks for POST and PUT, added confirmation request for DeleteByID and DeleteAll methods. Started work on the frontend.

+++ src/main/java/com/jamiemcmenemy/taskflow/controller/TaskController.java
@@ -4,18 +4,12 @@ import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.jamiemcmenemy.taskflow.dto.TaskDto;
 import com.jamiemcmenemy.taskflow.service.TaskService;
 
+import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
@@ -26,12 +20,12 @@ public class TaskController {
     private TaskService taskService;
 
     @PostMapping
-    public ResponseEntity<TaskDto> createTask(@RequestBody TaskDto taskDto){
+    public ResponseEntity<TaskDto> createTask(@Valid @RequestBody TaskDto taskDto){
         TaskDto createdTask = taskService.createTask(taskDto);
         return new ResponseEntity<>(createdTask, HttpStatus.CREATED);
     }
 
-    @GetMapping("{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<TaskDto> getTaskById(@PathVariable("id") Long taskId){
         TaskDto foundTask = taskService.getTaskByID(taskId);
         return ResponseEntity.ok(foundTask);
@@ -43,19 +37,29 @@ public class TaskController {
         return ResponseEntity.ok(tasks);
     }
 
-    @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteTask(@PathVariable("id") Long taskId){
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteTask(
+        @PathVariable("id") Long taskId,
+        @RequestParam boolean confirm){
+        if (!confirm) {
+            return ResponseEntity.badRequest().body("Please confirm deletion.");
+        }
         taskService.deleteTask(taskId);
-        return ResponseEntity.ok("Task" + taskId +"Deleted successfully");
+        return ResponseEntity.ok("Task " + taskId +" Deleted successfully");
     }
-    @DeleteMapping
-    public ResponseEntity<String> deleteAllTasks(){
+    @DeleteMapping()
+    public ResponseEntity<String> deleteAllTasks(
+        @RequestParam boolean confirm
+    ){
+        if (!confirm) {
+            return ResponseEntity.badRequest().body("Please confirm deletion.");
+        }
         taskService.deleteAllTasks();
-        return ResponseEntity.ok("Deleted all tasks succesfully.");
+        return ResponseEntity.ok("Deleted all tasks successfully.");
     }
 
-    @PutMapping("{id}")
-    public ResponseEntity<TaskDto> updateTask(@PathVariable("id") Long taskId, @RequestBody TaskDto updatedTask){
+    @PutMapping("/{id}")
+    public ResponseEntity<TaskDto> updateTask(@PathVariable("id") Long taskId, @Valid @RequestBody TaskDto updatedTask){
         TaskDto taskDto = taskService.updateTask(taskId, updatedTask);
         return ResponseEntity.ok(taskDto);
     }
